Split out doc upload tests to their own spec file
