+++ integration_tests/integration/book-recall.spec.js
@@ -424,193 +424,6 @@ context('Book a recall', () => {
     })
   })
 
-  it('User sees an uploaded document listed', () => {
-    const uploadDocuments = uploadDocumentsPage.verifyOnPage({ nomsNumber, recallId })
-    const documentId = '123'
-    cy.task('expectGetRecall', {
-      expectedResult: {
-        recallId,
-        ...getRecallResponse,
-        documents: [
-          {
-            category: 'UNCATEGORISED',
-            fileName: 'test.pdf',
-            documentId,
-          },
-        ],
-      },
-    })
-    uploadDocuments.upload({
-      filePath: '../test.pdf',
-      mimeType: 'application/pdf',
-    })
-    uploadDocuments.assertElementHasText({
-      qaAttr: `link-${documentId}`,
-      textToFind: 'test.pdf',
-    })
-    uploadDocuments.assertSelectValue({
-      fieldName: `category-${documentId}`,
-      value: 'UNCATEGORISED',
-    })
-  })
-
-  it('User sees previously saved documents', () => {
-    const documentId = '3fa85f64-5717-4562-b3fc-2c963f66afa6'
-    cy.task('expectGetRecall', {
-      expectedResult: {
-        recallId,
-        documents: [
-          {
-            category: 'PREVIOUS_CONVICTIONS_SHEET',
-            documentId,
-          },
-        ],
-      },
-    })
-    const uploadDocuments = uploadDocumentsPage.verifyOnPage({ nomsNumber, recallId })
-    uploadDocuments.assertElementHasText({
-      qaAttr: `link-${documentId}`,
-      textToFind: 'Pre Cons.pdf',
-    })
-    uploadDocuments.assertSelectValue({
-      fieldName: `category-${documentId}`,
-      value: 'PREVIOUS_CONVICTIONS_SHEET',
-    })
-  })
-
-  it('User sees an error if an upload fails', () => {
-    cy.task('expectAddRecallDocument', { statusCode: 400 })
-    const uploadDocuments = uploadDocumentsPage.verifyOnPage({ nomsNumber, recallId })
-    uploadDocuments.uploadSingleFile({
-      filePath: '../test.pdf',
-      mimeType: 'application/pdf',
-    })
-    uploadDocuments.assertSummaryErrorMessage({
-      fieldName: 'documents',
-      summaryError: 'test.pdf could not be uploaded - try again',
-    })
-  })
-
-  it('User sees an error if an upload has a virus', () => {
-    cy.task('expectAddRecallDocument', {
-      statusCode: 400,
-      responseBody: { status: 'BAD_REQUEST', message: 'VirusFoundException' },
-    })
-    const uploadDocuments = uploadDocumentsPage.verifyOnPage({ nomsNumber, recallId })
-    uploadDocuments.uploadSingleFile({
-      filePath: '../test.pdf',
-      mimeType: 'application/pdf',
-    })
-    uploadDocuments.assertSummaryErrorMessage({
-      fieldName: 'documents',
-      summaryError: 'test.pdf contains a virus',
-    })
-  })
-
-  it('user can go back to delete documents from the check your answers page', () => {
-    const documentId = '123'
-    cy.task('expectGetRecall', {
-      expectedResult: {
-        recallId,
-        ...getRecallResponse,
-        status: RecallResponse.status.BEING_BOOKED_ON,
-        documents: [
-          {
-            category: 'PART_A_RECALL_REPORT',
-            documentId,
-          },
-        ],
-      },
-    })
-    cy.task('expectDeleteRecallDocument')
-    const checkAnswers = checkAnswersPage.verifyOnPage({ nomsNumber, recallId })
-    checkAnswers.clickElement({ qaAttr: 'uploadedDocument-PART_A_RECALL_REPORT-Change' })
-    const uploadDocuments = uploadDocumentsPage.verifyOnPage()
-    uploadDocuments.clickElement({ qaAttr: `delete-${documentId}` })
-    uploadDocuments.clickContinue()
-    recallMissingDocumentsPage.verifyOnPage()
-  })
-
-  it("user can't go back to delete documents from the view recall page", () => {
-    const documentId = '123'
-    cy.task('expectGetRecall', {
-      expectedResult: {
-        recallId,
-        ...getRecallResponse,
-        status: RecallResponse.status.BEING_BOOKED_ON,
-        documents: [
-          {
-            category: 'PART_A_RECALL_REPORT',
-            documentId,
-          },
-        ],
-      },
-    })
-    cy.task('expectDeleteRecallDocument')
-    const recallInformation = recallInformationPage.verifyOnPage({ nomsNumber, recallId, personName })
-    recallInformation.clickElement({ qaAttr: 'uploadedDocument-PART_A_RECALL_REPORT-Change' })
-    const uploadDocuments = uploadDocumentsPage.verifyOnPage()
-    uploadDocuments.assertElementNotPresent({ qaAttr: `delete-${documentId}` })
-  })
-
-  it("user can't change the category of a document that has more than one version", () => {
-    const documentId = '123'
-    const recallResponse = {
-      recallId,
-      ...getRecallResponse,
-      status: RecallResponse.status.BEING_BOOKED_ON,
-      documents: [
-        {
-          category: 'PART_A_RECALL_REPORT',
-          documentId,
-          version: 2,
-        },
-      ],
-    }
-    cy.task('expectGetRecall', {
-      expectedResult: recallResponse,
-    })
-    cy.task('expectDeleteRecallDocument')
-    const recallInformation = recallInformationPage.verifyOnPage({ nomsNumber, recallId, personName })
-    recallInformation.clickElement({ qaAttr: 'uploadedDocument-PART_A_RECALL_REPORT-Change' })
-    const uploadDocuments = uploadDocumentsPage.verifyOnPage()
-    uploadDocuments.assertElementHasText({
-      qaAttr: `category-label-PART_A_RECALL_REPORT`,
-      textToFind: 'Part A recall report',
-    })
-    cy.task('expectGetRecall', {
-      expectedResult: {
-        ...recallResponse,
-        documents: [
-          ...recallResponse.documents,
-          {
-            category: 'PRE_SENTENCING_REPORT',
-            documentId: '456',
-          },
-        ],
-      },
-    })
-    uploadDocuments.uploadSingleFile({
-      filePath: '../uploads/presentencing_report.pdf',
-      mimeType: 'application/pdf',
-    })
-    uploadDocuments.assertElementPresent({ qaAttr: 'category-index-PRE_SENTENCING_REPORT' })
-  })
-
-  it("User sees an error if they don't upload the missing documents email or provide detail", () => {
-    cy.task('expectGetRecall', { recallId, expectedResult: { ...getEmptyRecallResponse, recallId } })
-    const recallMissingDocuments = recallMissingDocumentsPage.verifyOnPage({ nomsNumber, recallId })
-    recallMissingDocuments.clickContinue()
-    recallMissingDocuments.assertErrorMessage({
-      fieldName: 'missingDocumentsEmailFileName',
-      summaryError: 'Select an email',
-    })
-    recallMissingDocuments.assertErrorMessage({
-      fieldName: 'missingDocumentsDetail',
-      summaryError: 'Provide more detail',
-    })
-  })
-
   it('user can check and change their answers then navigate back to the check answers page', () => {
     cy.task('expectGetRecall', {
       expectedResult: { recallId, ...getRecallResponse, status: RecallResponse.status.BEING_BOOKED_ON },
