+++ integration_tests/integration/view-recall.spec.js
@@ -1,24 +1,13 @@
 import path from 'path'
-import { getEmptyRecallResponse, getRecallResponse, searchResponse } from '../mockApis/mockResponses'
-import { RecallResponse } from '../../server/@types/manage-recalls-api/models/RecallResponse'
-import uploadDocumentVersionPage from '../pages/uploadNewDocumentVersion'
+import { getEmptyRecallResponse, getRecallResponse } from '../mockApis/mockResponses'
 
 const recallInformationPage = require('../pages/recallInformation')
 
-context('View recall', () => {
+context('View a recall', () => {
   beforeEach(() => {
     cy.task('reset')
     cy.task('stubLogin')
     cy.task('stubAuthUser')
-    cy.task('expectListRecalls', {
-      expectedResults: [
-        {
-          recallId,
-          nomsNumber,
-        },
-      ],
-    })
-    cy.task('expectSearchResults', { expectedSearchTerm: nomsNumber, expectedSearchResults: searchResponse })
     cy.task('expectGetRecall', {
       recallId,
       expectedResult: {
@@ -79,6 +68,7 @@ context('View recall', () => {
       },
     })
     cy.task('expectGetUserDetails', { firstName: 'Bobby', lastName: 'Badger' })
+    cy.login()
   })
 
   const nomsNumber = 'A1234AA'
@@ -86,7 +76,6 @@ context('View recall', () => {
   const personName = 'Bobby Badger'
 
   it('User can view all recall information', () => {
-    cy.login()
     const recallInformation = recallInformationPage.verifyOnPage({ nomsNumber, recallId, personName })
     recallInformation.assertElementHasText({
       qaAttr: 'recallNotificationEmailSentDateTime',
@@ -151,7 +140,6 @@ context('View recall', () => {
         recallId,
       },
     })
-    cy.login()
     const recallInformation = recallInformationPage.verifyOnPage({ nomsNumber, recallId, personName })
     recallInformation.assertElementHasText({ qaAttr: 'additionalLicenceConditions', textToFind: 'Not available' })
     recallInformation.assertElementHasText({ qaAttr: 'vulnerabilityDiversity', textToFind: 'Not available' })
@@ -169,104 +157,13 @@ context('View recall', () => {
         contraband: false,
       },
     })
-    cy.login()
     const recallInformation = recallInformationPage.verifyOnPage({ nomsNumber, recallId, personName })
     recallInformation.assertElementHasText({ qaAttr: 'additionalLicenceConditions', textToFind: 'None' })
     recallInformation.assertElementHasText({ qaAttr: 'vulnerabilityDiversity', textToFind: 'No' })
     recallInformation.assertElementHasText({ qaAttr: 'contraband', textToFind: 'No' })
   })
 
-  it('user can go back from the view recall info page to add a new document version', () => {
-    const documentId = '123'
-    cy.task('expectGetRecall', {
-      expectedResult: {
-        recallId,
-        ...getRecallResponse,
-        status: RecallResponse.status.DOSSIER_IN_PROGRESS,
-        documents: [
-          {
-            category: 'PART_A_RECALL_REPORT',
-            documentId,
-            version: 2,
-            createdDateTime: '2021-11-21T12:34:30.000Z',
-          },
-        ],
-      },
-    })
-    cy.task('expectAddRecallDocument', { statusCode: 201 })
-    cy.login()
-    let recallInformation = recallInformationPage.verifyOnPage({ nomsNumber, recallId, personName })
-    recallInformation.clickElement({ qaAttr: 'uploadedDocument-PART_A_RECALL_REPORT-Change' })
-    const uploadDocumentVersion = uploadDocumentVersionPage.verifyOnPage({
-      documentCategoryLabel: 'part A recall report',
-    })
-    uploadDocumentVersion.assertLinkHref({
-      qaAttr: 'previousVersionFileName',
-      href: '/persons/A1234AA/recalls/123/documents/123',
-    })
-    uploadDocumentVersion.assertElementHasText({
-      qaAttr: 'previousVersionUploadedDateTime',
-      textToFind: 'Uploaded on 21 November 2021 at 12:34',
-    })
-    uploadDocumentVersion.uploadSingleFile({
-      filePath: '../test.pdf',
-      mimeType: 'application/pdf',
-    })
-    cy.task('expectGetRecall', {
-      expectedResult: {
-        recallId,
-        ...getRecallResponse,
-        status: RecallResponse.status.DOSSIER_IN_PROGRESS,
-        documents: [
-          {
-            category: 'PART_A_RECALL_REPORT',
-            documentId: '34589',
-            version: 2,
-          },
-        ],
-      },
-    })
-    uploadDocumentVersion.clickContinue()
-    recallInformation = recallInformationPage.verifyOnPage({ personName })
-    recallInformation.assertElementHasText({
-      qaAttr: 'uploadedDocument-PART_A_RECALL_REPORT-version',
-      textToFind: 'version 2',
-    })
-  })
-
-  it("user sees an error if they don't upload a new document version", () => {
-    const documentId = '123'
-    cy.task('expectGetRecall', {
-      expectedResult: {
-        recallId,
-        ...getRecallResponse,
-        status: RecallResponse.status.DOSSIER_IN_PROGRESS,
-        documents: [
-          {
-            category: 'PART_A_RECALL_REPORT',
-            documentId,
-            version: 2,
-            createdDateTime: '2021-11-21T12:34:30.000Z',
-          },
-        ],
-      },
-    })
-    cy.login()
-    const uploadDocumentVersion = uploadDocumentVersionPage.verifyOnPage({
-      nomsNumber,
-      recallId,
-      documentCategoryLabel: 'part A recall report',
-      documentCategoryName: 'PART_A_RECALL_REPORT',
-    })
-    uploadDocumentVersion.clickContinue()
-    uploadDocumentVersion.assertErrorMessage({
-      fieldName: 'documents',
-      summaryError: 'Select a file',
-    })
-  })
-
-  it('user can download all files', () => {
-    cy.login()
+  it('user can download all uploaded emails', () => {
     recallInformationPage.verifyOnPage({ nomsNumber, recallId, personName })
     const mockFileDownload = ({ fileName, docCategory }) => {
       cy.task('expectGetRecallDocument', {
